feat: added support for react-native + wasm
