+++ package.json
@@ -13,9 +13,9 @@
     "private set intersection",
     "private join and compute"
   ],
-  "main": "combined/wasm/cjs/index.js",
-  "browser": "combined/wasm/iife/index.js",
-  "module": "combined/wasm/es/index.js",
+  "main": "combined/wasm/node/cjs/index.js",
+  "browser": "combined/wasm/web/iife/index.js",
+  "module": "combined/wasm/web/es/index.js",
   "files": [
     "**/*.js"
   ],
@@ -39,7 +39,7 @@
     "build:proto": "bash private_set_intersection/javascript/scripts/build-proto.sh",
     "postbuild": "node -r esm private_set_intersection/javascript/version.js",
     "prerollup": "bash private_set_intersection/javascript/scripts/prerollup.sh",
-    "rollup": "rollup -c",
+    "rollup": "NODE_OPTIONS=--max_old_space_size=16384 rollup -c",
     "postrollup": "bash private_set_intersection/javascript/scripts/postrollup.sh",
     "prepublishOnly": "echo 'Please run \"npm run publish\" instead' && exit 1",
     "publish:test": "npm --dry-run publish --access=public private_set_intersection/javascript/dist/openmined-psi.js-*.tgz",
@@ -56,32 +56,32 @@
     "em:init": "bash private_set_intersection/javascript/scripts/em-init.sh"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.5",
-    "@babel/preset-env": "^7.10.4",
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
     "@rollup/plugin-alias": "^3.1.1",
-    "@rollup/plugin-commonjs": "^14.0.0",
-    "@rollup/plugin-node-resolve": "^8.4.0",
-    "@rollup/plugin-typescript": "^5.0.2",
-    "@types/google-protobuf": "^3.7.2",
-    "@types/jest": "^26.0.7",
-    "@typescript-eslint/eslint-plugin": "^3.7.0",
-    "@typescript-eslint/parser": "^3.7.0",
+    "@rollup/plugin-commonjs": "^15.1.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@rollup/plugin-typescript": "^6.0.0",
+    "@types/google-protobuf": "^3.7.3",
+    "@types/jest": "^26.0.14",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
     "buffer": "^5.6.0",
-    "eslint": "^7.5.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-jest": "^23.18.0",
-    "eslint-plugin-jsdoc": "^30.0.3",
+    "eslint": "^7.10.0",
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-jest": "^24.0.2",
+    "eslint-plugin-jsdoc": "^30.6.3",
     "eslint-plugin-prettier": "^3.1.4",
     "esm": "^3.2.25",
-    "jest": "^26.1.0",
+    "jest": "^26.4.2",
     "path": "^0.12.7",
-    "prettier": "^2.0.5",
-    "rollup": "^2.23.0",
+    "prettier": "^2.1.2",
+    "rollup": "^2.28.2",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-node-globals": "^1.4.0",
-    "rollup-plugin-terser": "^6.1.0",
-    "ts-jest": "^26.1.3",
-    "ts-protoc-gen": "^0.12.0",
-    "typescript": "^3.9.7"
+    "rollup-plugin-terser": "^7.0.2",
+    "ts-jest": "^26.4.1",
+    "ts-protoc-gen": "^0.13.0",
+    "typescript": "^4.0.3"
   }
 }
