+++ private_set_intersection/javascript/scripts/build-client-js.sh
@@ -5,4 +5,6 @@ set -euo pipefail
 source ./third_party/emsdk/emsdk_env.sh
 
 # Build the JS variant
-bazel build -c opt //private_set_intersection/javascript:psi_client_js.js --config=js
+bazel build -c opt //private_set_intersection/javascript:psi_client_js_node.js --config=js
+bazel build -c opt //private_set_intersection/javascript:psi_client_js_web.js --config=js
+bazel build -c opt //private_set_intersection/javascript:psi_client_js_worker.js --config=js
