+++ private_set_intersection/javascript/scripts/build-combined-js.sh
@@ -5,4 +5,6 @@ set -euo pipefail
 source ./third_party/emsdk/emsdk_env.sh
 
 # Build the JS variant
-bazel build -c opt //private_set_intersection/javascript:psi_combined_js.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_combined_js_node.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_combined_js_web.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_combined_js_worker.js --config=wasm
