+++ private_set_intersection/javascript/scripts/build-combined-wasm.sh
@@ -5,4 +5,6 @@ set -euo pipefail
 source ./third_party/emsdk/emsdk_env.sh
 
 # Build the WASM variant
-bazel build -c opt //private_set_intersection/javascript:psi_combined_wasm.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_combined_wasm_node.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_combined_wasm_web.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_combined_wasm_worker.js --config=wasm
