+++ private_set_intersection/javascript/scripts/build-server-wasm.sh
@@ -5,4 +5,6 @@ set -euo pipefail
 source ./third_party/emsdk/emsdk_env.sh
 
 # Build the WASM variant
-bazel build -c opt //private_set_intersection/javascript:psi_server_wasm.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_server_wasm_node.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_server_wasm_web.js --config=wasm
+bazel build -c opt //private_set_intersection/javascript:psi_server_wasm_worker.js --config=wasm
