+++ private_set_intersection/javascript/src/__tests__/client.test.ts
@@ -1,4 +1,4 @@
-import PSI from '../index_combined_wasm'
+import PSI from '../index_combined_wasm_node'
 import { ERROR_INSTANCE_DELETED } from '../implementation/constants'
 import { PSILibrary } from 'src/implementation/psi'
 import { Response, ServerSetup } from '../implementation/proto/psi_pb'
