+++ private_set_intersection/javascript/src/html/index_js.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8">
-        <script src="../../dist/combined/js/iife/index.js"></script>
-    </head>
-    <body>
-        You're debugging the JS target
-    </body>
+  <head>
+    <meta charset="utf-8" />
+    <script src="../../dist/combined/js/web/iife/index.js"></script>
+  </head>
+  <body>
+    You're debugging the JS target
+  </body>
 </html>
