+++ private_set_intersection/javascript/src/html/index_wasm.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8">
-        <script src="../../dist/combined/wasm/iife/index.js"></script>
-    </head>
-    <body>
-        You're debugging the WASM target
-    </body>
+  <head>
+    <meta charset="utf-8" />
+    <script src="../../dist/combined/wasm/web/iife/index.js"></script>
+  </head>
+  <body>
+    You're debugging the WASM target
+  </body>
 </html>
