+++ private_set_intersection/javascript/src/index_combined_js_web.ts
@@ -1,11 +1,12 @@
-import clientJsLibrary from 'psi_client_js'
+import combinedJsLibrary from 'psi_combined_js_web'
 
 import { NestedLibrary, createLoader } from './loader'
 import { PSILibrary, PSIConstructor } from './implementation/psi'
 import { PackageWrapperConstructor } from './implementation/package'
 import { ClientWrapperConstructor } from './implementation/client'
+import { ServerWrapperConstructor } from './implementation/server'
 
-const Loader = (): Promise<NestedLibrary> => createLoader(clientJsLibrary)
+const Loader = (): Promise<NestedLibrary> => createLoader(combinedJsLibrary)
 
 /**
  * Main export for the library
@@ -13,6 +14,9 @@ const Loader = (): Promise<NestedLibrary> => createLoader(clientJsLibrary)
 export default async (): Promise<PSILibrary> =>
   PSIConstructor({
     packageWrapper: PackageWrapperConstructor({
+      loader: await Loader()
+    }),
+    serverWrapper: ServerWrapperConstructor({
       loader: await Loader()
     }),
     clientWrapper: ClientWrapperConstructor({
