+++ private_set_intersection/javascript/src/index_combined_js_worker.ts
@@ -1,11 +1,12 @@
-import clientWasmLibrary from 'psi_client_wasm'
+import combinedJsLibrary from 'psi_combined_js_worker'
 
 import { NestedLibrary, createLoader } from './loader'
 import { PSILibrary, PSIConstructor } from './implementation/psi'
 import { PackageWrapperConstructor } from './implementation/package'
 import { ClientWrapperConstructor } from './implementation/client'
+import { ServerWrapperConstructor } from './implementation/server'
 
-const Loader = (): Promise<NestedLibrary> => createLoader(clientWasmLibrary)
+const Loader = (): Promise<NestedLibrary> => createLoader(combinedJsLibrary)
 
 /**
  * Main export for the library
@@ -13,6 +14,9 @@ const Loader = (): Promise<NestedLibrary> => createLoader(clientWasmLibrary)
 export default async (): Promise<PSILibrary> =>
   PSIConstructor({
     packageWrapper: PackageWrapperConstructor({
+      loader: await Loader()
+    }),
+    serverWrapper: ServerWrapperConstructor({
       loader: await Loader()
     }),
     clientWrapper: ClientWrapperConstructor({
