+++ private_set_intersection/javascript/src/index_combined_wasm_node.ts
@@ -1,11 +1,12 @@
-import serverJsLibrary from 'psi_server_js'
+import combinedWasmLibrary from 'psi_combined_wasm_node'
 
 import { NestedLibrary, createLoader } from './loader'
 import { PSILibrary, PSIConstructor } from './implementation/psi'
 import { PackageWrapperConstructor } from './implementation/package'
+import { ClientWrapperConstructor } from './implementation/client'
 import { ServerWrapperConstructor } from './implementation/server'
 
-const Loader = (): Promise<NestedLibrary> => createLoader(serverJsLibrary)
+const Loader = (): Promise<NestedLibrary> => createLoader(combinedWasmLibrary)
 
 /**
  * Main export for the library
@@ -16,6 +17,9 @@ export default async (): Promise<PSILibrary> =>
       loader: await Loader()
     }),
     serverWrapper: ServerWrapperConstructor({
+      loader: await Loader()
+    }),
+    clientWrapper: ClientWrapperConstructor({
       loader: await Loader()
     })
   })
