+++ private_set_intersection/javascript/src/index_combined_wasm_worker.ts
@@ -1,0 +1,25 @@
+import combinedWasmLibrary from 'psi_combined_wasm_worker'
+
+import { NestedLibrary, createLoader } from './loader'
+import { PSILibrary, PSIConstructor } from './implementation/psi'
+import { PackageWrapperConstructor } from './implementation/package'
+import { ClientWrapperConstructor } from './implementation/client'
+import { ServerWrapperConstructor } from './implementation/server'
+
+const Loader = (): Promise<NestedLibrary> => createLoader(combinedWasmLibrary)
+
+/**
+ * Main export for the library
+ */
+export default async (): Promise<PSILibrary> =>
+  PSIConstructor({
+    packageWrapper: PackageWrapperConstructor({
+      loader: await Loader()
+    }),
+    serverWrapper: ServerWrapperConstructor({
+      loader: await Loader()
+    }),
+    clientWrapper: ClientWrapperConstructor({
+      loader: await Loader()
+    })
+  })
