+++ private_set_intersection/javascript/src/index_server_wasm_worker.ts
@@ -1,12 +1,11 @@
-import combinedJsLibrary from 'psi_combined_js'
+import serverWasmLibrary from 'psi_server_wasm_worker'
 
 import { NestedLibrary, createLoader } from './loader'
 import { PSILibrary, PSIConstructor } from './implementation/psi'
 import { PackageWrapperConstructor } from './implementation/package'
-import { ClientWrapperConstructor } from './implementation/client'
 import { ServerWrapperConstructor } from './implementation/server'
 
-const Loader = (): Promise<NestedLibrary> => createLoader(combinedJsLibrary)
+const Loader = (): Promise<NestedLibrary> => createLoader(serverWasmLibrary)
 
 /**
  * Main export for the library
@@ -17,9 +16,6 @@ export default async (): Promise<PSILibrary> =>
       loader: await Loader()
     }),
     serverWrapper: ServerWrapperConstructor({
-      loader: await Loader()
-    }),
-    clientWrapper: ClientWrapperConstructor({
       loader: await Loader()
     })
   })
