+++ private_set_intersection/javascript/toolchain/cc_toolchain_config.bzl
@@ -111,7 +111,7 @@ def _impl(ctx):
         # Overriding to use -O2 instead of -O3 because asmjs breaks.
         flag_set(
             actions = all_compile_actions + all_link_actions,
-            flag_groups = [flag_group(flags = ["-g0", "-O2"])], 
+            flag_groups = [flag_group(flags = ["-g0", "-O2"])],
             with_features = [with_feature_set(features = ["opt"])],
         ),
         # Fastbuild (fastbuild)
